moving login page to '/' , moving home page to '/home', bug fixes, optimization
