+++ src/app/(app)/home/job_card.js
@@ -6,7 +6,7 @@ import {Rubik} from "next/font/google";
 import {ErrorOutline} from "@mui/icons-material";
 import Button from "@mui/joy/Button";
 import {useRouter} from "next/navigation";
-import HomeViewStatusChanger from "@/app/(app)/(home)/status_change";
+import HomeViewStatusChanger from "@/app/(app)/home/status_change";
 
 const rubik = Rubik({subsets: ['latin']})
