+++ src/app/(app)/job/page.js
@@ -7,6 +7,6 @@ export default function Page() {
     const router = useRouter()
 
     useEffect(() => {
-        router.push('/')
+        router.push('/home')
     })
 }
